+++ .gitignore
@@ -15,6 +15,7 @@ downloads/
 eggs/
 .eggs/
 lib/
+data/
 lib64/
 parts/
 sdist/
@@ -129,3 +130,5 @@ dmypy.json
 .pyre/
 
 .vscode
+exomiser/
+exomiser*.zip
