+++ pheval/db.py
@@ -4,18 +4,16 @@
 from __future__ import generators
 import jaydebeapi
 import pandas as pd
-import random
 from .decorators import *
 import logging as log
 from tqdm import tqdm
+from randomisation import *
 
 log.basicConfig(
     level=log.INFO,
     format="%(asctime)-8s %(levelname)-8s %(message)s",
     datefmt="%Y-%m-%d",
 )
-
-random.seed(10)
 
 
 class DBConnector:
@@ -128,45 +126,6 @@ def ResultIterator(query: str, conn: DBConnection, arraysize: int = 1000) -> tup
             yield result
 
 
-def rand(
-    self: pd.DataFrame, min_num: int, max_num: int, scrambe_factor: float
-) -> float:
-    """
-    Numeric scrambling
-    Args:
-        self (pd.DataFrame): dataframe records
-        min_num (int): min value from this records
-        max_num (int): max value from this records
-        scrambe_factor (float): scramble factor scalar
-
-    Returns:
-        float: randomized number
-    """
-    try:
-        return self + (random.uniform(min_num, max_num) * scrambe_factor)
-    except Exception as err:
-        log.error(self, exc_info=err)
-        return self
-
-
-def ssp_randomisation(df, scramble_factor=0.5) -> pd.DataFrame:
-    """
-      Takes as input a pandas dataframe with the semantic similarity profile and a value between 0 and 1 (scramble_factor: 0 means no scrambling and 1 means  complete randomisation). It then randomises the above scores with a degree of the scramble_factor and returns a scrambles pandas dataframe.
-    Args:
-          df (pd.DataFrame): original records as dataframe came from phenotype database
-          scramble_factor (float) scalar scramble factor
-
-    Returns:
-        pd.Dataframe: scrambled dataframe
-    """
-    columns = ["SIMJ", "IC", "SCORE"]
-    for c in columns:
-        min_num = df[c].min()
-        max_num = df[c].max()
-        df[c] = df[c].apply(rand, args=(min_num, max_num, scramble_factor))
-        return df
-
-
 def create_table(conn: DBConnection, table_name: str):
     """
     Creating table
@@ -259,7 +218,7 @@ def select(
     return data_to_update
 
 
-@memory(percentage=0.5)
+@memory(percentage=0.95)
 def scramble_table(table_name: str, scramble_factor: float) -> None:
     with connector as cnn:
         conn = DBConnection(cnn)
@@ -269,7 +228,7 @@ def scramble_table(table_name: str, scramble_factor: float) -> None:
         log.info(
             f"Scrambling records from table: {table_name} using {scramble_factor} magnitude"
         )
-        insert(conn, table_name, scramble_factor, count_original, 300000)
+        insert(conn, table_name, scramble_factor, count_original, 3000)
         count_scramble = count_total(conn, f"{table_name}_SCRAMBLE")
         log.info(f"Scrambled records length: {count_scramble}")
         log.info("Done")
